--- a/crates/swc_ecma_minifier/Cargo.toml
+++ b/crates/swc_ecma_minifier/Cargo.toml
@@ -40,7 +40,7 @@ swc_ecma_utils = {version = "0.69.0", path = "../swc_ecma_utils"}
 swc_ecma_visit = {version = "0.54.0", path = "../swc_ecma_visit"}
 swc_timer = {version = "0.4.0", path = "../swc_timer"}
 tracing = "0.1.31"
-unicode-xid = "0.2.2"
+unicode-id = "0.3.0"
 
 [dev-dependencies]
 ansi_term = "0.12.1"