--- a/crates/swc_ecma_parser/src/lexer/util.rs
+++ b/crates/swc_ecma_parser/src/lexer/util.rs
@@ -1,8 +1,5 @@
 //! Ported from [babylon/util/identifier.js][]
 //!
-//! Note: Currently this use xid instead of id because unicode_xid crate
-//! exists.
-//!
 //!
 //! [babylon/util/identifier.js]:https://github.com/babel/babel/blob/master/packages/babylon/src/util/identifier.js
 use std::char;