--- a/crates/swc_ecma_parser/tests/typescript/issue-3862/index.ts
+++ b/crates/swc_ecma_parser/tests/typescript/issue-3862/index.ts
@@ -1,0 +1,5 @@
+var ゛; // U+309B
+var \u309B;
+var \u309C;
+var a\u309B;
+var a\u309C;