… Cargo.lock
… crates/swc_ecma_parser/tests/typescript/issue-3341/input.ts