@Boshen 2022-05-20
I published a ~2x faster and 2kb smaller https://crates.io/crates/unicode-id-start, a fork from https://github.com/dtolnay/unicode-ident, would you like to give it a try and bench? Tests are passing in the Rome repo https://github.com/rome/tools/pull/2600