@kdy1 2022-05-20
@Boshen It seems to be actually slower.

master:
```
     Running benches/parser.rs (/Users/kdy1/projects/master-swc/target/release/deps/parser-3fa8a5c343758a1f)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
es/parser/colors        time:   [14.627 us 14.653 us 14.695 us]
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) high mild
  8 (8.00%) high severe

es/parser/angular       time:   [7.6256 ms 7.6393 ms 7.6583 ms]
Found 11 outliers among 100 measurements (11.00%)
  11 (11.00%) high severe

Benchmarking es/parser/backbone: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.3s, enable flat sampling, or reduce sample count to 60.
es/parser/backbone      time:   [1.0485 ms 1.0501 ms 1.0520 ms]
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe

es/parser/jquery        time:   [5.8608 ms 5.8670 ms 5.8741 ms]
Found 16 outliers among 100 measurements (16.00%)
  9 (9.00%) high mild
  7 (7.00%) high severe

es/parser/jquery mobile time:   [9.2385 ms 9.2490 ms 9.2612 ms]
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe

es/parser/mootools      time:   [4.5426 ms 4.5555 ms 4.5703 ms]
Found 16 outliers among 100 measurements (16.00%)
  16 (16.00%) high severe

es/parser/underscore    time:   [912.10 us 913.79 us 916.06 us]
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe

es/parser/three         time:   [26.538 ms 26.596 ms 26.674 ms]
Found 17 outliers among 100 measurements (17.00%)
  6 (6.00%) high mild
  11 (11.00%) high severe

es/parser/yui           time:   [4.5326 ms 4.5425 ms 4.5558 ms]
Found 16 outliers among 100 measurements (16.00%)
  5 (5.00%) high mild
  11 (11.00%) high severe


```


With `unocide-id-start`


```
    Finished bench [optimized + debuginfo] target(s) in 47.80s
     Running benches/parser.rs (/Users/kdy1/projects/s/perf/target/release/deps/parser-fc34b825e591d37b)
WARNING: HTML report generation will become a non-default optional feature in Criterion.rs 0.4.0.
This feature is being moved to cargo-criterion (https://github.com/bheisler/cargo-criterion) and will be optional in a future version of Criterion.rs. To silence this warning, either switch to cargo-criterion or enable the 'html_reports' feature in your Cargo.toml.

Gnuplot not found, using plotters backend
es/parser/colors        time:   [14.964 us 14.988 us 15.013 us]
                        change: [-0.1938% +0.0311% +0.2661%] (p = 0.79 > 0.05)
                        No change in performance detected.
Found 23 outliers among 100 measurements (23.00%)
  7 (7.00%) low severe
  10 (10.00%) low mild
  1 (1.00%) high mild
  5 (5.00%) high severe

es/parser/angular       time:   [7.6198 ms 7.6294 ms 7.6413 ms]
                        change: [-1.5204% -1.2412% -0.9750%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  2 (2.00%) high mild
  6 (6.00%) high severe

Benchmarking es/parser/backbone: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.3s, enable flat sampling, or reduce sample count to 60.
es/parser/backbone      time:   [1.0530 ms 1.0573 ms 1.0619 ms]
                        change: [-1.9814% -1.2038% -0.4412%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe

es/parser/jquery        time:   [5.8711 ms 5.8894 ms 5.9114 ms]
                        change: [-1.2975% -0.9091% -0.4743%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  5 (5.00%) high mild
  9 (9.00%) high severe

es/parser/jquery mobile time:   [9.2244 ms 9.2401 ms 9.2619 ms]
                        change: [-1.6064% -1.3953% -1.1664%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe

es/parser/mootools      time:   [4.5215 ms 4.5279 ms 4.5372 ms]
                        change: [-1.6521% -1.4016% -1.1575%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) high mild
  11 (11.00%) high severe

es/parser/underscore    time:   [915.10 us 916.52 us 918.25 us]
                        change: [-1.3710% -1.0421% -0.6754%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  3 (3.00%) high mild
  12 (12.00%) high severe

es/parser/three         time:   [26.553 ms 26.596 ms 26.647 ms]
                        change: [-1.0552% -0.7822% -0.5174%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 18 outliers among 100 measurements (18.00%)
  6 (6.00%) high mild
  12 (12.00%) high severe

es/parser/yui           time:   [4.5356 ms 4.5415 ms 4.5485 ms]
                        change: [-0.8592% -0.6833% -0.5203%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  9 (9.00%) high mild
  6 (6.00%) high severe

```