@Boshen 2022-05-20
Try removing the ascii chars check? The crate handles it with a lookup table https://github.com/Boshen/unicode-id-start/blob/master/src/tables.rs#L9-L21